refactor: move rpc state reader to state reader crate
commit-id:c590a350
