+++ crates/gateway/src/lib.rs
@@ -4,10 +4,6 @@ mod compiler_version;
 pub mod config;
 pub mod errors;
 pub mod gateway;
-mod rpc_objects;
-mod rpc_state_reader;
-#[cfg(test)]
-mod rpc_state_reader_test;
 
 mod stateful_transaction_validator;
 mod stateless_transaction_validator;
