+++ crates/mempool_node/Cargo.toml
@@ -19,6 +19,7 @@ starknet_gateway = { path = "../gateway", version = "0.0" }
 starknet_mempool = { path = "../mempool", version = "0.0" }
 starknet_mempool_infra = { path = "../mempool_infra", version = "0.0" }
 starknet_mempool_types = { path = "../mempool_types", version = "0.0" }
+starknet_state_reader = { path = "../state_reader", version = "0.0" }
 tokio.workspace = true
 tracing.workspace = true
 validator.workspace = true
@@ -27,6 +28,6 @@ validator.workspace = true
 assert-json-diff.workspace = true
 assert_matches.workspace = true
 colored.workspace = true
+mempool_test_utils = { path = "../mempool_test_utils" }
 pretty_assertions.workspace = true
 serde_json.workspace = true
-mempool_test_utils = { path = "../mempool_test_utils" }
