+++ crates/state_reader/Cargo.toml
@@ -12,7 +12,20 @@ workspace = true
 testing = []
 
 [dependencies]
+axum.workspace = true
 blockifier = { workspace = true, features = ["testing"] }
+cairo-lang-starknet-classes.workspace = true
 mempool_test_utils = { path = "../mempool_test_utils", version = "0.0" }
+papyrus_config.workspace = true
+papyrus_rpc.workspace = true
+reqwest.workspace = true
+serde.workspace = true
+serde_json.workspace = true
 starknet-types-core.workspace = true
 starknet_api.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+validator.workspace = true
+
+[dev-dependencies]
+mockito.workspace = true
