+++ crates/state_reader/src/config.rs
@@ -1,0 +1,33 @@
+use std::collections::BTreeMap;
+
+use papyrus_config::dumping::{ser_param, SerializeConfig};
+use papyrus_config::{ParamPath, ParamPrivacyInput, SerializedParam};
+use serde::{Deserialize, Serialize};
+use validator::Validate;
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize, Validate, PartialEq)]
+pub struct RpcStateReaderConfig {
+    pub url: String,
+    pub json_rpc_version: String,
+}
+
+#[cfg(any(feature = "testing", test))]
+impl RpcStateReaderConfig {
+    pub fn create_for_testing() -> Self {
+        Self { url: "http://localhost:8080".to_string(), json_rpc_version: "2.0".to_string() }
+    }
+}
+
+impl SerializeConfig for RpcStateReaderConfig {
+    fn dump(&self) -> BTreeMap<ParamPath, SerializedParam> {
+        BTreeMap::from_iter([
+            ser_param("url", &self.url, "The url of the rpc server.", ParamPrivacyInput::Public),
+            ser_param(
+                "json_rpc_version",
+                &self.json_rpc_version,
+                "The json rpc version.",
+                ParamPrivacyInput::Public,
+            ),
+        ])
+    }
+}
