+++ crates/state_reader/src/lib.rs
@@ -1,4 +1,7 @@
+pub mod config;
+pub mod errors;
+mod rpc_objects;
+pub mod rpc_state_reader;
 pub mod state_reader;
-
 #[cfg(any(feature = "testing", test))]
 pub mod state_reader_test_utils;
