+++ crates/state_reader/src/rpc_state_reader.rs
@@ -1,3 +1,7 @@
+#[cfg(test)]
+#[path = "rpc_state_reader_test.rs"]
+mod rpc_state_reader_test;
+
 use blockifier::blockifier::block::BlockInfo;
 use blockifier::execution::contract_class::{ContractClass, ContractClassV0, ContractClassV1};
 use blockifier::state::errors::StateError;
@@ -9,7 +13,6 @@ use serde_json::{json, Value};
 use starknet_api::block::BlockNumber;
 use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce};
 use starknet_api::state::StorageKey;
-use starknet_state_reader::state_reader::{MempoolStateReader, StateReaderFactory};
 use starknet_types_core::felt::Felt;
 
 use crate::config::RpcStateReaderConfig;
@@ -19,6 +22,7 @@ use crate::rpc_objects::{
     GetCompiledContractClassParams, GetNonceParams, GetStorageAtParams, RpcResponse,
     RPC_CLASS_HASH_NOT_FOUND, RPC_ERROR_BLOCK_NOT_FOUND, RPC_ERROR_CONTRACT_ADDRESS_NOT_FOUND,
 };
+use crate::state_reader::{MempoolStateReader, StateReaderFactory};
 
 pub struct RpcStateReader {
     pub config: RpcStateReaderConfig,
